Adds basic items (hats)
Had to do some major refactoring for this and discovered several
limitations of the architecture, but managed to get it to work.

 * QueryableAggregator now inherits Component. This way a Component of an
Entity can also be a QueryableAggregator and have components of it's own.
  * This is important because the elements belonging to an Item component
don't have to be untangled from a Character that Item component is
attached to
 * Adds EquipableComponent (spelling?)
  * Each EquipableComponent corresponds to a slot
 * Adds InteractionService
  * Selects the default interaction when an entity enters a new tile
  * Equips any equipable item in that square
 * Renames CharacterComponent to NameComponent (it occurred that of course
non-character entities need names, too)
 * Adds a HatGenerator for generating the all-important hats

GET ABOARD THE HAT COLLECTING SIMULATOR 2018 TRAIN CHOO CHOO